+++ fractions/microprofile/microprofile-health/src/main/java/org/wildfly/swarm/health/HealthMetaData.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.monitor;
+package org.wildfly.swarm.health;
 
 /**
  * @author Heiko Braun
