+++ fractions/microprofile/microprofile-health/src/main/java/org/wildfly/swarm/health/deployment/HealthResponseFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.monitor.deployment;
+package org.wildfly.swarm.health.deployment;
 
 import org.wildfly.swarm.health.Status;
