+++ fractions/microprofile/microprofile-health/src/main/java/org/wildfly/swarm/health/runtime/BufferingSinkConduit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.monitor.runtime;
+package org.wildfly.swarm.health.runtime;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
