+++ fractions/microprofile/microprofile-health/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-org.wildfly.swarm.monitor.deployment.HealthExtension
+org.wildfly.swarm.health.deployment.HealthExtension
