+++ fractions/microprofile/microprofile-health/src/main/resources/modules/org/eclipse/microprofile/health/main/module.xml
@@ -8,7 +8,7 @@
     <dependencies>
         <module name="javax.enterprise.api"/>
         <module name="org.wildfly.swarm.undertow" slot="main"/>
-        <module name="org.wildfly.swarm.monitor" slot="main" services="import"/>
+        <!--<module name="org.wildfly.swarm.monitor" slot="main" services="import"/>-->
         <module name="org.eclipse.microprofile.config.api" export="true"/>
         <module name="org.wildfly.extension.microprofile.config" export="true"/>
     </dependencies>
